delete fonctionnel

+++ app/Http/Controllers/PostController.php
@@ -89,7 +89,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post): RedirectResponse
     {
-        // $this->authorize("delete", $post);
+        $this->authorize("delete", $post);
 
         $post->delete();
